+++ source/STM32firmware/PlusCart/Src/cartridge_emulation_df.c
@@ -40,7 +40,7 @@ void emulate_dfsc_cartridge(const char* filename, uint32_t image_size, uint8_t* 
 				while (ADDR_IN == addr) { data_prev = data; data = DATA_IN; }
 				data = data_prev;
 
-				ram[address] = data;
+				ram[address] = (uint8_t) data;
 			} else {
 				if (address >= 0x0fc0 && address <= 0x0fdf) bank = layout->banks[address - 0x0fc0];
